requirements.txt
influxdb3-python
pandas
tabulate
influxdb-client